Modify heartrate to healthdata (#30)
* :recycle: Refactor HeartRate Route/Handler/Service/Adapter/Port.

At before, I planned to create each Route/Handler/Service of each healthdata.
But the number of kind of healthdatas is more than 100, so it could introduces some overhead.

Because of that, I'm planning to  create only one Router, Handler, Service, Adapter, Port for all healthdatas.
This PR is related with this perspective.


* :sparkles: Remove findById and create findByPeriod.


* :fix: fix ktlintcheck error


* :recycle: Modify CreateHealthDataCommand and fix omitted heartrate updating

CreateHealthDataCommand should be able to accept all kinds of HealthData types.
At CreateHealthDataService, the service will map a proper domain by checking 'type'.
All HealthData types have userId. Excepting userId, fields can be different among HealthDatas.
So fields will be stored at 'data'. And the service will map this datas to proper field, too.


* :recycle: Reflect reviewers' requests.

Add HealthDataType enum class on HealthData class.
Add HealthDataRepository. This will be parent class of all HealthData types' repository.
Add HealthDataEntity. This will be parent class of all HealthData entities.


* :bug: Rollback HealthDataRepository update.

* :recycle: Rename GetHealthDataUseCase to GetHealthDataQuery

+++ platform/src/main/resources/db/migration/V0.0.3__create_heartrates_table.sql
@@ -1,7 +1,7 @@
 create TABLE "heartrates"
 (
     "id"         SERIAL      NOT NULL,
-    "user_id"    SERIAL      NOT NULL,
+    "user_id"    INT8      NOT NULL,
     "time"       timestamptz NOT NULL,
     "bpm"        INT8     NOT NULL,
     PRIMARY KEY ("id")
